requirements.txt
imutils==0.5.4
numpy==1.22.4
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
pytesseract==0.3.9
